slurm-mcp/pyproject.toml
[project]
name = "slurm-mcp"
version = "0.1.0"
description = "Slurm MCP server implementation using Model Context Protocol"
requires-python = ">=3.10"
dependencies = [
  "mcp[cli]>=0.1.0",
  "pytest-asyncio>=1.0.0",
  "python-dotenv>=1.0.0",
  "psutil>=5.9.0",
]

[project.optional-dependencies]
test = ["pytest>=7.0.0"]

[project.scripts]
slurm-mcp = "server:main"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[dependency-groups]
dev = [
    "pytest>=8.4.0",
]

[tool.uv.workspace]
members = [
    "slurm-mcp",
]
